requirements.txt
pytest==8.3.2
requests==2.32.3
pydantic==2.9.2
Faker==30.1.0
python-dotenv==1.0.1
locust==2.32.3
